A bunch of changes to adjust log box, combat screen and old html message box to my laptop res; made smooth move happen at 60 fps and synched it with whatever the author sets for realTime timer length; made smooth, but non-realtime default; fixed that finally toolset changes for smooth move and realTime false/true are correctly applied; improved how player perceives collision with a moving prop before encounter/convo happens; fixed portrait loading (redundand now); tiles were drawn only partially in combat (fixed; strange  - I should have noticed that on normal desktop pc, too, confused here); right hand orientated party and enemies in combat were drawn half a sqaure too much to the right side, corrected (again, strange I did not see that on normal desktop pc)

+++ IceBlink2/ScreenMainMap.cs
@@ -551,12 +551,13 @@ namespace IceBlink2
 
                             int tileNumberOfPropSquare = SquareThatPixIsOnX + (SquareThatPixIsOnY * gv.mod.currentArea.MapSizeX);
 
-                            int pixDistanceOfPropToPlayerX = (p.currentPixelPositionX - playerPositionXInPix);
+                            //cast the pix position to int in order to draw it at nearly exact loc
+                            int pixDistanceOfPropToPlayerX = ((int)p.currentPixelPositionX - playerPositionXInPix);
                             if (pixDistanceOfPropToPlayerX < 0)
                             {
                                 pixDistanceOfPropToPlayerX *= -1;
                             }
-                            int pixDistanceOfPropToPlayerY = (p.currentPixelPositionY - playerPositionYInPix);
+                            int pixDistanceOfPropToPlayerY = ((int)p.currentPixelPositionY - playerPositionYInPix);
                             if (pixDistanceOfPropToPlayerY < 0)
                             {
                                 pixDistanceOfPropToPlayerY *= -1;
@@ -564,7 +565,7 @@ namespace IceBlink2
 
                             if ((pixDistanceOfPropToPlayerX <= ((gv.playerOffset + 1) * gv.squareSize)) && (pixDistanceOfPropToPlayerY <= ((gv.playerOffset + 1) * gv.squareSize)))
                             {
-                                IbRect dst = new IbRect(p.currentPixelPositionX, p.currentPixelPositionY, gv.squareSize, gv.squareSize);
+                                IbRect dst = new IbRect((int)p.currentPixelPositionX, (int)p.currentPixelPositionY, gv.squareSize, gv.squareSize);
                                 gv.DrawBitmap(p.token, src, dst);
 
                                 if (mod.showInteractionState == true)
